Visual representation of ascii board
