--- frontend/src/features/chessboard/chessboard.ts
@@ -1,1 +1,0 @@
-import { Chess } from 'chess.js'
